# Devirtualize UTF8 with a specialized text converter