--- a/src/Npgsql/Internal/ResolverFactories/LTreeTypeInfoResolverFactory.cs
+++ b/src/Npgsql/Internal/ResolverFactories/LTreeTypeInfoResolverFactory.cs
@@ -31,15 +31,15 @@ sealed class LTreeTypeInfoResolverFactory : PgTypeInfoResolverFactory
         {
             mappings.AddType<string>("ltree",
                 static (options, mapping, _) => mapping.CreateInfo(options,
-                    new VersionPrefixedTextConverter<string>(LTreeVersion, new StringTextConverter(options.TextEncoding))),
+                    new VersionPrefixedTextConverter<string>(LTreeVersion, StringTextConverter.Create(options.TextEncoding))),
                 MatchRequirement.DataTypeName);
             mappings.AddType<string>("lquery",
                 static (options, mapping, _) => mapping.CreateInfo(options,
-                    new VersionPrefixedTextConverter<string>(LTreeVersion, new StringTextConverter(options.TextEncoding))),
+                    new VersionPrefixedTextConverter<string>(LTreeVersion, StringTextConverter.Create(options.TextEncoding))),
                 MatchRequirement.DataTypeName);
             mappings.AddType<string>("ltxtquery",
                 static (options, mapping, _) => mapping.CreateInfo(options,
-                    new VersionPrefixedTextConverter<string>(LTreeVersion, new StringTextConverter(options.TextEncoding))),
+                    new VersionPrefixedTextConverter<string>(LTreeVersion, StringTextConverter.Create(options.TextEncoding))),
                 MatchRequirement.DataTypeName);
 
             return mappings;