--- a/test/Npgsql.Benchmarks/TypeHandlers/Text.cs
+++ b/test/Npgsql.Benchmarks/TypeHandlers/Text.cs
@@ -1,12 +1,12 @@
 ﻿using BenchmarkDotNet.Attributes;
-using System.Collections.Generic;
-using System.Text;
+using Npgsql.Internal;
 using Npgsql.Internal.Converters;
+using System.Collections.Generic;
 
 namespace Npgsql.Benchmarks.TypeHandlers;
 
 [Config(typeof(Config))]
-public class Text() : TypeHandlerBenchmarks<string>(new StringTextConverter(Encoding.UTF8))
+public class Text() : TypeHandlerBenchmarks<string>(StringTextConverter.Create(NpgsqlWriteBuffer.UTF8Encoding))
 {
     protected override IEnumerable<string> ValuesOverride()
     {