`Encoding.UTF8` actually returns a UTF8EncodingSealed object (https://github.com/dotnet/runtime/blob/main/src/libraries/System.Private.CoreLib/src/System/Text/UTF8Encoding.Sealed.cs) that allows for the functions to be devirtualized if accessed via that static.

Npgsql is creating its own UTF8Encoding objects that don't allow for this specialization.

Using a custom UTF8Encoding and a specific TextConverter when that encoding is in use allows greater performance. I did not see much performance differences with the rest of the stack allocations in UTF8EncodingSealed.

Before
| Method | Value                | Mean      | Error    | StdDev   | Op/s         | Gen0   | Allocated |
|------- |--------------------- |----------:|---------:|---------:|-------------:|-------:|----------:|
| Read   | x                    |  24.72 ns | 0.098 ns | 0.087 ns | 40,449,395.2 | 0.0023 |      24 B |
| Write  | x                    |        NA |       NA |       NA |           NA |     NA |        NA |
| Read   | xxxxxxxxxx           |  26.19 ns | 0.078 ns | 0.069 ns | 38,187,844.1 | 0.0046 |      48 B |
| Write  | xxxxxxxxxx           |  18.96 ns | 0.027 ns | 0.024 ns | 52,748,503.0 | 0.0000 |         - |
| Read   | xxxx(...)xxxx [100]  |  35.01 ns | 0.649 ns | 0.575 ns | 28,564,979.3 | 0.0214 |     224 B |
| Write  | xxxx(...)xxxx [100]  |  24.38 ns | 0.071 ns | 0.059 ns | 41,019,998.9 | 0.0005 |       5 B |
| Read   | xxxx(...)xxxx [1000] | 107.83 ns | 0.627 ns | 0.556 ns |  9,273,447.3 | 0.1935 |    2024 B |
| Write  | xxxx(...)xxxx [1000] |  71.47 ns | 0.165 ns | 0.147 ns | 13,992,508.8 | 0.0051 |      54 B |

After
| Method | Value                | Mean     | Error    | StdDev   | Op/s         | Gen0   | Allocated |
|------- |--------------------- |---------:|---------:|---------:|-------------:|-------:|----------:|
| Read   | x                    | 18.88 ns | 0.161 ns | 0.135 ns | 52,956,082.8 | 0.0023 |      24 B |
| Write  | x                    |       NA |       NA |       NA |           NA |     NA |        NA |
| Read   | xxxxxxxxxx           | 20.33 ns | 0.093 ns | 0.078 ns | 49,191,801.5 | 0.0046 |      48 B |
| Write  | xxxxxxxxxx           | 18.83 ns | 0.019 ns | 0.015 ns | 53,103,167.5 | 0.0000 |         - |
| Read   | xxxx(...)xxxx [100]  | 27.33 ns | 0.091 ns | 0.085 ns | 36,587,970.4 | 0.0214 |     224 B |
| Write  | xxxx(...)xxxx [100]  | 21.84 ns | 0.028 ns | 0.023 ns | 45,783,841.6 | 0.0005 |       5 B |
| Read   | xxxx(...)xxxx [1000] | 98.62 ns | 0.773 ns | 0.723 ns | 10,139,873.0 | 0.1935 |    2024 B |
| Write  | xxxx(...)xxxx [1000] | 67.47 ns | 0.227 ns | 0.201 ns | 14,822,340.7 | 0.0051 |      54 B |