@bbowyersmyth 2025-02-08
@NinoFloris is this a change you would be interested in? I know there is some desire for small AOT size.