@NinoFloris 2025-02-08
Clever observation! Maybe we can devirtualize inside Read and Write? We would add a branch `if (Encoding.UTF8 == encoding)` but won't need the extra converter and so on.