@NinoFloris 2025-02-08
Ah I missed the fact Npgsql actually defines its own UTF8Encoding instance as well. Since we don't call GetPreamble anywhere we can just use Encoding.UTF8 for NpgsqlWriteBuffer.UTF8Encoding. However we still need to use Encoding.UTF8 directly in the converter for the JIT to see it's of type UTF8EncodingSealed.