@bbowyersmyth 2025-02-09
`Encoding.UTF8` uses the "replace invalid chars" option whereas the main Npgsql encoding uses the "throw on invalid" setting. 

I tried to combine the converter but most things were a regression on the original perf. Had to deconstruct the Read function and then it was good. The current change might be a bit worse for AOT but the JIT likes it.
Only modifying the sync Read at the moment.

| Method | Value                | Mean     | Error    | StdDev   | Op/s         | Gen0   | Allocated |
|------- |--------------------- |---------:|---------:|---------:|-------------:|-------:|----------:|
| Read   | x                    | 16.68 ns | 0.066 ns | 0.058 ns | 59,945,087.0 | 0.0023 |      24 B |
| Write  | x                    |       NA |       NA |       NA |           NA |     NA |        NA |
| Read   | xxxxxxxxxx           | 18.09 ns | 0.073 ns | 0.068 ns | 55,264,866.3 | 0.0046 |      48 B |
| Write  | xxxxxxxxxx           | 19.68 ns | 0.028 ns | 0.025 ns | 50,810,920.2 | 0.0000 |         - |
| Read   | xxxx(...)xxxx [100]  | 25.24 ns | 0.079 ns | 0.074 ns | 39,616,495.6 | 0.0214 |     224 B |
| Write  | xxxx(...)xxxx [100]  | 24.95 ns | 0.053 ns | 0.050 ns | 40,080,983.7 | 0.0005 |       5 B |
| Read   | xxxx(...)xxxx [1000] | 98.19 ns | 1.330 ns | 1.244 ns | 10,184,553.7 | 0.1935 |    2024 B |
| Write  | xxxx(...)xxxx [1000] | 72.26 ns | 0.242 ns | 0.214 ns | 13,838,010.8 | 0.0051 |      54 B |