@NinoFloris 2025-02-09
>Encoding.UTF8 uses the "replace invalid chars" option whereas the main Npgsql encoding uses the "throw on invalid" setting.

I traced back this decision to https://github.com/npgsql/npgsql/issues/503, however it seems outdated in the current scheme of things. At that time no other encodings were supported (support was introduced in ea9f203713f2cabad5b1e12b78d542827a4eb1b3). So we're currently in the situation where we only do the strict thing for UTF8, which I don't think makes a lot of sense.

I'd personally would be fine using Encoding.UTF8 throughout, @roji what are your thoughts on this?